Fail quietly for incompatible PHP syntax files
Fixes #5.  The Hack syntax file is written to augment the default PHP
syntax file that ships with Vim 7.4.  This note has been added to the
README, and the Hack syntax file now defines all syntax groups it
references before using them so that it doesn't scream too loudly if
some other syntax file is used.