syntax: allow whitespace before a generic
Hack allows for whitespace between an identifier and a generic type. For
instance, the following are the same:

```
function getFooBar<T>(MyGeneric<T> $foo): string {
function getFooBar<T>(MyGeneric <T> $foo): string {
```

Without this change, the above would cause the ` <T>` to be interpreted
as XHP, and would mess up highlighting for the rest of the file.